research_api/requirements/base.txt
Django>=5.1,<7.0
djangorestframework>=3.15
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
spacy>=3.7,<4.0
scikit-learn>=1.4
numpy>=1.26
httpx>=0.27
